positive-ev/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
SQLAlchemy==2.0.27
pandas==2.2.0
plotly==5.18.0
selenium==4.18.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
python-dotenv==1.0.1
numpy==1.26.4
requests==2.31.0
python-crontab==3.0.0
ratelimit==2.2.1
openai==1.12.0
google-api-python-client==2.118.0 